# docs/openapi.json
{
  "components": {
    "schemas": {
      "AuditEventRead": {
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "event_type"
        ],
        "title": "AuditEventRead",
        "type": "object"
      },
      "FileImportCreate": {
        "properties": {
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          },
          "simulate_failure": {
            "anyOf": [
              {
                "enum": [
                  "retryable",
                  "permanent"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulate Failure"
          },
          "source_format": {
            "default": "json",
            "enum": [
              "json",
              "csv"
            ],
            "title": "Source Format",
            "type": "string"
          },
          "source_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Source Name",
            "type": "string"
          }
        },
        "required": [
          "source_name",
          "records"
        ],
        "title": "FileImportCreate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MembershipCreate": {
        "properties": {
          "role": {
            "enum": [
              "owner",
              "admin",
              "manager",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "maxLength": 36,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "role"
        ],
        "title": "MembershipCreate",
        "type": "object"
      },
      "MembershipRead": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "manager",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "role",
          "status"
        ],
        "title": "MembershipRead",
        "type": "object"
      },
      "OutboxEventRead": {
        "properties": {
          "adapter_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Key"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dead_lettered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dead Lettered At"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Duration Ms"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "next_retry_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Retry At"
          },
          "payload_json": {
            "title": "Payload Json",
            "type": "string"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "result_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Json"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "event_type",
          "payload_json",
          "status",
          "attempts"
        ],
        "title": "OutboxEventRead",
        "type": "object"
      },
      "TenantCreate": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 80,
            "minLength": 2,
            "pattern": "^[a-z0-9][a-z0-9-]*[a-z0-9]$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "TenantCreate",
        "type": "object"
      },
      "TenantRead": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "status"
        ],
        "title": "TenantRead",
        "type": "object"
      },
...
      "UserRead": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "status"
        ],
        "title": "UserRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DriveDesk Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "operationId": "list_tenants_endpoint_tenants_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantRead"
                  },
                  "title": "Response List Tenants Endpoint Tenants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tenants Endpoint"
      },
      "post": {
        "operationId": "create_tenant_endpoint_tenants_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tenant Endpoint"
      }
    },
    "/tenants/{tenant_id}": {
      "get": {
        "operationId": "get_tenant_endpoint_tenants__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tenant Endpoint"
      }
    },
    "/tenants/{tenant_id}/audit-events": {
      "get": {
        "operationId": "list_audit_events_endpoint_tenants__tenant_id__audit_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventRead"
                  },
                  "title": "Response List Audit Events Endpoint Tenants  Tenant Id  Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events Endpoint"
      }
    },
    "/tenants/{tenant_id}/integration-imports/file": {
      "post": {
        "operationId": "create_file_import_endpoint_tenants__tenant_id__integration_imports_file_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileImportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboxEventRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create File Import Endpoint"
      }
    },
    "/tenants/{tenant_id}/memberships": {
      "get": {
        "operationId": "list_memberships_endpoint_tenants__tenant_id__memberships_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MembershipRead"
                  },
                  "title": "Response List Memberships Endpoint Tenants  Tenant Id  Memberships Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memberships Endpoint"
      },
      "post": {
        "operationId": "create_membership_endpoint_tenants__tenant_id__memberships_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Membership Endpoint"
      }
    },
    "/tenants/{tenant_id}/outbox-events": {
      "get": {
        "operationId": "list_outbox_events_endpoint_tenants__tenant_id__outbox_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutboxEventRead"
                  },
                  "title": "Response List Outbox Events Endpoint Tenants  Tenant Id  Outbox Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Outbox Events Endpoint"
      }
    },
    "/users": {
      "get": {
        "operationId": "list_users_endpoint_users_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "title": "Response List Users Endpoint Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users Endpoint"
      },
      "post": {
        "operationId": "create_user_endpoint_users_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "system",
              "title": "X-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "viewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User Endpoint"
      }
    }
  }
}